test(tools): consolidate file edit flow coverage
